Add absorbance helper function
